src/ZNI/requirements.txt
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dj-rest-auth==2.2.5
Django==4.1.1
djangorestframework==3.13.1
idna==3.4
importmagic3==0.2.0
jedi==0.18.1
log-colorizer==2.0.0
parso==0.8.3
psutil==5.9.2
python-magic==0.4.27
pytz==2022.2.1
requests==2.28.1
six==1.16.0
spark-parser==1.8.9
sqlparse==0.4.2
tornado==5.1.1
tornado-systemd==1.0.1
tzdata==2022.2
uncompyle6==3.8.0
urllib3==1.26.12
wdb==3.3.0
wdb.server==3.3.0
xdis==6.0.4
